integration-tests/requirements.txt
kafka-python==2.0.2
requests==2.31.0
black==23.3.0
ruff==0.0.270
mypy==1.3.0
types-requests==2.31.0.1